+++ webextension/js/init-templates.js
@@ -14,7 +14,6 @@ export function nunjuckRender(templateName, context) {
 const templatesSource = window.templatesSource = new Map();
 templatesSource.set('backgroundTheme', '/templates/backgroundTheme');
 templatesSource.set('tabMover', '/templates/panel/tabMover');
-templatesSource.set('tabPageServerIp', '/templates/panel/tabPageServerIp');
 templatesSource.set('newTab', '/templates/newTab');